aria-bot/requirements.txt
discord.py==2.3.2
anthropic==0.25.0
aiosqlite==0.20.0
python-dotenv==1.0.1
aiohttp==3.9.5
